bot/requirements.txt
discord.py[voice]==2.7.1
davey==0.1.4
git+https://github.com/rdphillips7/discord-ext-voice-recv.git@6b9fb95f4b23dbf17f0e32c9f94bef822252794d
deepgram-sdk>=3.5.0
supabase==2.10.0
PyNaCl==1.5.0
python-dotenv==1.0.1
structlog==24.4.0
